Place `INFO` command implementation in separate command (#1216)
